+++ src/components/NavbarButtons.vue
@@ -11,7 +11,7 @@ const { isDarkTheme } = toRefs(styleStore);
     <c-button
       circle
       variant="text"
-      href="https://github.com/CorentinTh/it-tools"
+      href="https://github.com/archiguru/it-tools-zh_CN"
       target="_blank"
       rel="noopener noreferrer"
       :aria-label="$t('home.nav.githubRepository')"
@@ -20,7 +20,7 @@ const { isDarkTheme } = toRefs(styleStore);
     </c-button>
   </c-tooltip>
 
-  <c-tooltip :tooltip="$t('home.nav.twitter')" position="bottom">
+  <!-- <c-tooltip :tooltip="$t('home.nav.twitter')" position="bottom">
     <c-button
       circle
       variant="text"
@@ -31,7 +31,7 @@ const { isDarkTheme } = toRefs(styleStore);
     >
       <n-icon size="25" :component="BrandTwitter" />
     </c-button>
-  </c-tooltip>
+  </c-tooltip> -->
 
   <c-tooltip :tooltip="$t('home.nav.about')" position="bottom">
     <c-button circle variant="text" to="/about" :aria-label="$t('home.nav.aboutLabel')">