+++ vite.config.ts
@@ -12,15 +12,17 @@ import { NaiveUiResolver } from 'unplugin-vue-components/resolvers';
 import Components from 'unplugin-vue-components/vite';
 import { defineConfig } from 'vite';
 import { VitePWA } from 'vite-plugin-pwa';
-import markdown from 'vite-plugin-vue-markdown';
+import Markdown from 'unplugin-vue-markdown/vite'
 import svgLoader from 'vite-svg-loader';
 import { configDefaults } from 'vitest/config';
+import { visualizer } from 'rollup-plugin-visualizer';
 
 const baseUrl = process.env.BASE_URL ?? '/';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
+    //@ts-ignore
     VueI18n({
       runtimeOnly: true,
       jitCompilation: true,
@@ -28,7 +30,7 @@ export default defineConfig({
       fullInstall: true,
       strictMessage: false,
       include: [
-        resolve(__dirname, 'locales/**'),
+        resolve(__dirname, 'locales/*.yml'),
       ],
     }),
     AutoImport({
@@ -47,11 +49,13 @@ export default defineConfig({
       },
     }),
     Icons({ compiler: 'vue3' }),
+    visualizer(),
     vue({
       include: [/\.vue$/, /\.md$/],
     }),
     vueJsx(),
-    markdown(),
+    //@ts-ignore
+    Markdown(),
     svgLoader(),
     VitePWA({
       registerType: 'autoUpdate',